BEET tool

+++ src/app/modules/tools/distribution-transformer/distribution-transformer.component.ts
@@ -1,13 +1,13 @@
-import { Component, OnInit , Inject, SystemJsNgModuleLoader} from '@angular/core';
+import { Component, OnInit, Inject, SystemJsNgModuleLoader } from '@angular/core';
 import { FormArray, FormBuilder, Validators, FormGroup } from '@angular/forms';
 import { dateInputsHaveChanged } from '@angular/material/datepicker/datepicker-input-base';
 import { ChartDataSets, ChartOptions, ChartType } from 'chart.js';
 import { Label, SingleDataSet } from 'ng2-charts';
-import { DialogData, DistTransformer} from 'src/app/shared/models/models';
+import { DialogData, DistTransformer } from 'src/app/shared/models/models';
 import { ToolsService } from 'src/app/shared/services/tools.service';
-import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { DialogOverviewExampleDialog } from './newtransformer-dialog/distribution-transformer-dialog';
-import {Location} from '@angular/common';
+import { Location } from '@angular/common';
 import { ConfirmationDialogService } from 'src/app/shared/services/confirmation-dialog.service';
 
 export interface FinalTable {
@@ -35,13 +35,13 @@ export interface PieArray {
   styleUrls: ['./distribution-transformer.component.scss']
 })
 export class DisTransformerComponent implements OnInit {
-  public pieChartOptions: ChartOptions = {responsive: true,};
+  public pieChartOptions: ChartOptions = { responsive: true, };
   pieChartLabels: Label[] = [];
   pieChartData: SingleDataSet = [];
   public pieChartType: ChartType = 'pie';
   public pieChartLegend = true;
   public pieChartPlugins = [];
-  barChartOptions: ChartOptions ={responsive: true,};
+  barChartOptions: ChartOptions = { responsive: true, };
   barChartLabels: Label[] = ['CO2 baseline', 'Final CO2 emissions'];
   barChartType: ChartType = 'bar';
   barChartLegend = true;
@@ -53,7 +53,7 @@ export class DisTransformerComponent implements OnInit {
   submitted: boolean;
   errMessage: string;
   dataSource: any;
-  inputTableDataSource : any ;
+  inputTableDataSource: any;
   countrylist: String[];
   highestVoltageValueslist: String[];
   dualVoltWindings: String[];
@@ -67,7 +67,7 @@ export class DisTransformerComponent implements OnInit {
     'FireRegulation',
     'HighestVoltageValues',
     'DualVoltWindings',
-    'Capex','action'
+    'Capex', 'action'
   ]
   displayedColumns: string[] = [
     'efficacy',
@@ -90,18 +90,18 @@ export class DisTransformerComponent implements OnInit {
   InputValuesArr: DistTransformer[] = [];
   inputValues: DistTransformer[] = [];
   ctr = 0;
-  
-  constructor(private fb: FormBuilder,
-     private toolService: ToolsService,
-     public dialog: MatDialog,
-     private _location: Location,
-     private confirmationDialog: ConfirmationDialogService) { 
 
-     }
+  constructor(private fb: FormBuilder,
+    private toolService: ToolsService,
+    public dialog: MatDialog,
+    private _location: Location,
+    private confirmationDialog: ConfirmationDialogService) {
+
+  }
 
   backClicked() {
-   location.replace(location.origin);
-  } 
+    location.replace(location.origin);
+  }
 
   ngOnInit(): void {
     this.toolService.getTool1Inputs().subscribe(res => {
@@ -117,28 +117,28 @@ export class DisTransformerComponent implements OnInit {
     });
   }
 
-  openDialog(action,obj): void {
+  openDialog(action, obj): void {
     const dialogRef = this.dialog.open(DialogOverviewExampleDialog, {
       width: '80%',
       autoFocus: false,
       maxHeight: '90vh',
       data: {
         highestVoltageValueslist: this.highestVoltageValueslist,
-        dualVoltWindings: this.dualVoltWindings ,
-        requestCount : this.inputValues.length,
-        action: action, 
+        dualVoltWindings: this.dualVoltWindings,
+        requestCount: this.inputValues.length,
+        action: action,
         selectedTransformer: obj
       }
     });
 
-   
+
 
     dialogRef.afterClosed().subscribe(result => {
       console.log(result);
-      if(result !== undefined){
-        this.inputValues= this.inputValues.filter(x => x.Requestnumber != result.Requestnumber)
+      if (result !== undefined) {
+        this.inputValues = this.inputValues.filter(x => x.Requestnumber != result.Requestnumber)
         this.inputValues.push(result);
-        this.inputValues= this.inputValues.sort((first, second) => 0 - (first.Requestnumber > second.Requestnumber ? -1 : 1));
+        this.inputValues = this.inputValues.sort((first, second) => 0 - (first.Requestnumber > second.Requestnumber ? -1 : 1));
         this.inputTableDataSource = this.inputValues;
       }
     });
@@ -146,19 +146,18 @@ export class DisTransformerComponent implements OnInit {
 
 
   public openConfirmationDialog(obj) {
-    this.confirmationDialog.confirm('Delete Transformer','Do you really want to delete?')
-    .then((confirmed) => this.deleteTransformer(confirmed,obj))
-    .catch(() => console.log('User dismissed the dialog'));
+    this.confirmationDialog.confirm('Delete Transformer', 'Do you really want to delete?')
+      .then((confirmed) => this.deleteTransformer(confirmed, obj))
+      .catch(() => console.log('User dismissed the dialog'));
   }
 
-  deleteTransformer(confirm,obj) : void {
+  deleteTransformer(confirm, obj): void {
     console.log(confirm);
-    if(confirm)
-    {
+    if (confirm) {
       this.inputValues = this.inputValues.filter(i => i.Requestnumber !== obj.Requestnumber)
       this.inputTableDataSource = this.inputValues;
     }
-  } 
+  }
 
   addNewTransformerVariant() {
     var transReq = {
@@ -168,7 +167,7 @@ export class DisTransformerComponent implements OnInit {
     }
 
 
-this.toolService.sendDisTransformerDetails(transReq).subscribe(res => {
+    this.toolService.sendDisTransformerDetails(transReq).subscribe(res => {
       console.log(res);
       if (res.status == 'success') {
         var finalOutValues: FinalTable[] = [];
@@ -200,17 +199,17 @@ this.toolService.sendDisTransformerDetails(transReq).subscribe(res => {
           this.pieChartLabels.push('E' + i);
           this.pieChartData.push(Math.round(calEnergy * 10) / 10);
         }
-      
+
         this.barChartData.push()
         console.log(this.dataSource, this.pieLabels, this.pieData);
         this.showForms = false;
-      } 
-      
+      }
+
       else {
         this.showForms = true;
       }
     });
 
-    
+
   }
 }
